fix: change payment flow
